requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.3.9
cffi==1.12.2
cryptography==2.6.1
future==0.17.1
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycparser==2.19
pylint==2.3.1
python-telegram-bot==11.1.0
six==1.12.0
trueskill==0.4.5
typed-ast==1.4.1
wrapt==1.11.1
